Revert "[Robustness] use `string.prototype.trim`"
This reverts commit f645b490bb5e223c567acf102fb9ae87f2b5e6c6, and fixes #9.

The rationale to keep the dependency is that more deps is better, and
robustness matters more than other considerations.

The rationale to remove it is that it significantly inflates bundle
size, in a non-major version (which only matters in that users are less
likely to audit the release and understand the consequences); and that
v1.0.0 for years lacked robustness around trimming, without complaint,
so the trim package may not even be necessary.